refactor: move ansible to the project dir
