+++ ansible/__init__.py
@@ -10,8 +10,8 @@ from typing import List
 import ansible_runner
 from ansible_runner import Runner, AnsibleRunnerException
 
-from deploy.ansible import ansible_const
-from deploy.ansible.exceptions import AnsibleExecuteError
+from ansible import ansible_const
+from ansible.exceptions import AnsibleExecuteError
 
 logger = logging.getLogger('ansible_deploy')
